+++ src/tools/tools.jl
@@ -101,10 +101,10 @@ julia> using PDBTools
 julia> atoms = read_pdb(PDBTools.SMALLPDB);
 
 julia> center_of_mass(atoms)
-3-element StaticArraysCore.SVector{3, Float64} with indices SOneTo(3):
-  -5.584422772707132
- -13.110413081059928
-  -7.139970851058855
+3-element StaticArraysCore.SVector{3, Float32} with indices SOneTo(3):
+  -5.5844226
+ -13.1104145
+  -7.1399713
 ```
 
 """
@@ -123,7 +123,7 @@ end
     atoms = [Atom(name="C", x=-1.0, y=-1.0, z=-1.0), Atom(name="C", x=1.0, y=1.0, z=1.0)]
     @test center_of_mass(atoms) ≈ [0.0, 0.0, 0.0] atol = 1e-10
     atoms = read_pdb(PDBTools.SMALLPDB)
-    @test center_of_mass(atoms) ≈ [-5.584422752942997, -13.110413157869903, -7.139970815730879] atol = 1e-7
+    @test center_of_mass(atoms) ≈ [-5.584422752942997, -13.110413157869903, -7.139970815730879] atol = 1e-5
 end
 
 """
@@ -139,18 +139,18 @@ julia> using PDBTools
 julia> atoms = read_pdb(PDBTools.SMALLPDB);
 
 julia> center_of_mass(atoms)
-3-element StaticArraysCore.SVector{3, Float64} with indices SOneTo(3):
-  -5.584422772707132
- -13.110413081059928
-  -7.139970851058855
+3-element StaticArraysCore.SVector{3, Float32} with indices SOneTo(3):
+  -5.5844226
+ -13.1104145
+  -7.1399713
 
 julia> moveto!(atoms; center = [1.0, 2.0, 3.0]);
 
 julia> center_of_mass(atoms)
-3-element StaticArraysCore.SVector{3, Float64} with indices SOneTo(3):
- 1.0000000263619948
- 1.9999999934852166
- 2.9999999509668918
+3-element StaticArraysCore.SVector{3, Float32} with indices SOneTo(3):
+ 1.0
+ 2.0000014
+ 3.0
 ```
 """
 function moveto!(atoms::AbstractVector{<:Atom}; center::AbstractVector{<:Real}=SVector(0.0, 0.0, 0.0))
@@ -166,6 +166,6 @@ end
 @testitem "moveto!" begin
     using PDBTools
     atoms = read_pdb(PDBTools.SMALLPDB)
-    @test center_of_mass(moveto!(atoms)) ≈ [0.0, 0.0, 0.0] atol = 1e-7
-    @test center_of_mass(moveto!(atoms; center=[1.0, 2.0, 3.0])) ≈ [1.0, 2.0, 3.0] atol = 1e-7
+    @test center_of_mass(moveto!(atoms)) ≈ [0.0, 0.0, 0.0] atol = 1e-5
+    @test center_of_mass(moveto!(atoms; center=[1.0, 2.0, 3.0])) ≈ [1.0, 2.0, 3.0] atol = 1e-5
 end